chore(index.js): clean code, move tool method into common
